+++ ui/packages/catalog-ui-search/src/main/webapp/component/result-item/result-item.view.js
@@ -17,7 +17,7 @@ import React from 'react'
 import MarionetteRegionContainer from '../../react-component/container/marionette-region-container'
 import styled from 'styled-components'
 import { readableColor } from 'polished'
-import MetacardInteractionsDropdown from '../../react-component/container/metacard-interactions/metacard-interactions-dropdown'
+import MetacardInteractionsDropdown from '../../react-component/metacard-interactions/metacard-interactions-dropdown'
 const Backbone = require('backbone')
 const Marionette = require('marionette')
 const CustomElements = require('../../js/CustomElements.js')
