+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/metacard-interactions/hide-interaction.tsx
@@ -13,9 +13,9 @@
  *
  **/
 import * as React from 'react'
-import { MetacardInteraction } from '../../presentation/metacard-interactions/metacard-interactions'
+import { MetacardInteraction } from './metacard-interactions'
 import { Props, Result, Model } from '.'
-const user = require('../../../component/singletons/user-instance')
+const user = require('../../component/singletons/user-instance')
 import { hot } from 'react-hot-loader'
 
 const handleShow = (props: Props) => {
